backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
aiohttp==3.9.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-nmap==0.7.1
beautifulsoup4==4.12.2
lxml==4.9.3
cryptography==41.0.7
pillow==10.1.0
python-multipart==0.0.6
httpx==0.25.2
openai==1.3.7
ollama==0.1.7
Jinja2==3.1.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1